// src/helpers/schema.tsx
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;
// ...
export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export type TitleColor = "blue" | "brown" | "green" | "orange" | "purple" | "rose" | "teal";
// ...
export interface MemberTitles {
  color: TitleColor;
  createdAt: Generated<Timestamp>;
  description: string | null;
  id: Generated<number>;
  isActive: Generated<boolean>;
  isDefault: Generated<boolean>;
  name: string;
}
// ...
export const TitleColorArrayValues: [TitleColor, ...TitleColor[]] = ["blue","brown","green","orange","purple","rose","teal"];